+++ subtest/tests/expand.rs
@@ -20,6 +20,7 @@ pub fn test_snapshots() {
 /// Also actually run the tests for which we do snapshot testing above
 mod expand {
     mod async_tests;
+    mod doc_comments;
     #[expect(
         dead_code,
         reason = "since #[test] is missing, the function should lead to a dead code warning, \